Add diffutils and gettext ports
There's a TODO on the gettext port.
We need the ability to limit run dependencies on specific packages.
This is seen by libprintf package which requires gcc runtime libs,
but the other packages do not.

as an absurd case, we don't want to pull in gcc for a docs package,
for example.