joe: fix on *BSD
Ever since ncurses was updated to version 6.1, joe has segfaulted
immediately on BSD.  The only solution I've found is to link it against
ncurses 6.0.  I've added a static-only ncurses60 port to accomplish this.
While it is intended as a build-only dependency, it will install
alongside the standard ncurses without conflict.