dmidecode: remove OS-specific patches (#95)
The dragonfly patch did not apply.  The freebsd patch was guarded by
macros, so there was no reason it shouldn't have been in /patches.
So the dragonfly bit was added to it.

Note that ports with os-specific patches are harder to maintain.
They should only be used if changes CONFLICT between platforms.
Otherwise, try to use macro guards and have one patch.