unibilium: switch to neovim fork
github/mauke/unibilium has been abandoned by the developer. As possibly
the only project that actually uses it, neovim has forked it and
continues development of it. This commit switches our port to that fork.

The original intent is to fix home/end keys in neovim under screen. Those
keys do not work in tmux even after this commit, so that will need some
attention.