Add PGplot (create cmake framework)
The reason it took so long to build this ports is that the
entire makefile system had to be rewritten.  FreeBSD did this
but relied on their own make system so it had to be rewritten
for Ravenports.  I used cmake.

Linux support may need a tweak.  Built on DragonFly.

While here, fix perl-PGPLOT