librsvg: upgrade
This requires rust which isn't available on all plaforms (notably
solaris right now).  To handle this, add a "RUSTLESS" option which
is off by default.  For those operating systems that don't have
rust, set this option on.  In those cases it will build the latest
version not requiring rust.

Note in this case, it will be packaged using the latest librsvg
release version, not the true version (e.g. 2.40.21).