Fix chicken-egg problem for bootstraps/platform upgrades
Currently bmake is getting built with values that most come
from the stored files inside of the sysroot (with the exception
of OPSYS and ARCH).  If the sysroot gets rebuilt as part of an
upgrade without a corresponding bmake rebuild (which has to happen
after the sysroot upgrade), the internal values of bmake are
wrong, causing havok down the road.  It's even a bigger issue for
bootstrapping on a new platform where bmake never existed before.

The solution is to hold the platform specifications centrally in
raven.version.mk and extract that information to build both
sysroot and bmake.  Additionally a check has been added to sysroot
to ensure bmake has been rebuilt with the new specifications.
If the packaged bmake doesn't match the new specifications, the
sysroot build fails fatally.

The corresponding change to bmake is coming an a future commit.