Erlang: fix build on FreeBSD
Erlang doesn't build with clang due to (known?) issue with a
couple of unreferenced symbols.

Erlang wouldn' build on FreeBSD either, tickling a long-standing
PCH bug on GCC on FreeBSD.  However, that's only used on ASMJIT
and GCC can build Erlang if we turn of JIT support.  So that's
what I did on FreeBSD.