Intentionally break network after fetch stage
We do not want any network activity after the fetching phase.
Until now, we could not enforce this policy.
Now, as extract starts, the resolv configuration will be changed
to point to localhost.  For most build machines, this will break
the ability to resolve DNS and break fetch.

Two things come from this:
1) Expect some ports to be broken as a result.
   (We want this so we can add the dependencies)
2) Ravenadm ENTERAFTER feature will add "fetch" as a valid
   stopping phase so we can stop with the network intact.

The original resolv.conf is preserved so it can be manually added
back as necessary.  Don't programmatically use that though!