python numpy: make it functional on netbsd (fixes pandas)
NetBSD doesn't provide expm1l, log1pl or log2l

I was able to use the builtin function for log2l but the same
approach did not work for the other two, so it was downgraded
to the double versions. So numpy will not be precise on those
functions but most of it works.