first draft at golang port generator (and upgrade git-lfs)
This might still need some work because I can't tell which modules
need extra reloc.  In this case all jcmturner files required that
the extra path be dropped and others required to keep it.  They
look the same to me on the graph.

I didn't mean to commit the git-lfs port with this, but the
git commit --amend command sucked them in.