rgbds: Remove rerolled note  (#262)
So it turns out that github naively changed the compression algorithm
on the generated tarballs and caused massive breakage across the
repositories of the world.  They quickly reverted the change after
outrage ensued.

Rgbds was never retagged; we were just running the build tests after
each change by github.  We're back to the original checksum.

Removing the reroll note as suggested by at
https://github.com/gbdev/rgbds/issues/1128