Drop perl-Data-Munge
The author released version "0.10" six years after "0.097".  Per version
convention, this should have been "0.100".  While "0.10" and "0.100" are
numerically equivalent (as floating point numbers), they are not
equivalent as "versions".  As a result, "0.10" is considered older than
"0.097".

The author refuses to re-release as "0.100", thinking the rest of the
world needs to work around him.

Luckily, since Data::Munge was introduced to Ravenports, all other
perl modules have dropped it as a dependency.  It is now an orphan.
Rather than fake the version, let's just drop the entire port.  It's
too insignificant to spend any more time arguing about it.

(Breaking link to avoid it getting posted to repology)
https:// github.com / repology / repology-rules / issues / 697