skia is only meant for linux, mac, and windows
It can build on BSD, but there are undefined symbols.
I tried changing some "is_linux" to "true" in the BUILD.gn file
and that fixed the SkDebugf issue, but another issue with piex
popped up.  It's going to require more patching so disable
everywhere except linux.

The port is only used by fonts-cantarell