Drop libadacrypt
It stopped building with gcc13.
It could have been fixed with relaxing checks.
However, checking upstream, this is project was last committed to 5 years
ago and a disclaimer added that it's basically a collaboration of
student projects:

Note that the ACL is an ongoing academic projects. Several parts of
the code is written by students as part of a semester project or a
(bachelor/master) thesis. Usually, those students are neither ADA
professionals nor cryptographic experts. Thus, use this project with
caution.

It's already a "unique" port in Repology.
Time to let it go.