Drop AdaControl
This port is one of two that requires ASIS.
ASIS was last buildable by GCC 9 Gnat compiler which is not gone.
This port was left in place to see if there was a solution to
restore the port with the latest FSF GCC but that was 2-3 years
ago so it is time to say goodbye. (Sorry Pierre)