Add new port ravensys-root-dragonfly
This serves as the template to split out system roots into their own
separate ports with a minimum of duplication.

This approach also allows the support of MULTIPLE system roots
per OS/Arch.  In this case, DragonFly 6.2 and 6.4.  While one
is always considered "standard", the non-standard variants can
be separately installed for use as a ravenadm profile.  This allows
for current support of multiple releases of the same OS/Arch
combination.