bmake: create variants to support older system roots
The make utility has operation system information encoded into it.
This information has to match the system root that it's designed
for.  The system root build actually checks that these match.

Three variants were created to support the older system roots
of FreeBSD, DragonFly and Linux.  NetBSD and Midnight currently
only support a single release so no variants for them.