ConsoleKit2: switch to arrowd branch
This branch was created to support FreeBSD.  It's main contribution
is systemd shims.  It broke the gtk-doc support though, so
documentation has been dropped.

The manifest has been simplified now that DragonFly produces the
same system scripts as the other platforms (by patching to treat
DragonFly as FreeBSD)

A couple of patches from FreeBSD have been brought in as well.
Finally, the X11 support has been converted to optional, but it
is still on by default.