ravensw: upgrade to solve several memory leaks
The big win here is the "checking integrity" crash is fixed.
This primarily manifested on Linux but it could show on any
platform during install.

Valgrind shows that ravensw (derived from pkg fork) fails to
deallocate lots of structures prior to exit.  For now I'm not
going to try to resolve those leaks because the program is
ending anyway.  Hopefully this will greatly improve the
stability of ravensw (c).