Add support for sccache when cargo is used.
It looks in two places for sccache:
1) /usr/bin/sccache
2) $LOCALBASE/bin/sccache

That's the search order.  When the program is found, it sets RUSTC_WRAPPER
to the path of the found sccache, and then sets SCCACHE_DIR to a
subdirectory of CCACHE_DIR.

If BUILD_WITH_CCACHE is not defined, sccache is disabled just as ccache is.

It's intended to add sccache to the system roots, which is going to add a
hugh build dependency to it, so it will probably set an an option.