rust: upgrade and finally fix https verification
A few months ago, after we fixed the crate.io https verification by
fixing openssl-probe, it broke again.  The openssl-probe changes were
still valid, so the breakage was a mystery.

That mystery was solved today.  Silently rust was changed to build
embedded curl by default instead of using the system libcurl.  So the
embedded curl didn't have the CURL_CA_BUNDLE defined, and for some
reason rust didn't use it's own openssl-probe, so it couldn't find
the certs.  The solution was to patch vendor/curl-sys to define the
location of the cert bundle.