webkit2: fix linking on MidnightBSD
I'm not sure what the problem was here.  The MOLD linker was being invoked
but the symptoms were the same as the ld.bfd linker.  The fix was to
add USES=gold (which luckily is added after USES=mold) which fixed the
linking.

I have built on FreeBSD yet so I'm not sure what FreeBSD is going to do
(in other words, is this both FBSD and Midnight, or just Midnight?)