python.mk: only mark when generating plists
It turns out that even if USES=python isn't set, it could still be brought
in, which forces QA check to look for include/ and lib/pkgconf files in
non-generated manifest ports that aren't even python (e.g. gtk-doc).

Move the mark to the autoplist target to ensure that QA check only
happens on python ports with generated manifests.