curl: rework while added 'embed' variant
During this I noticed the transmission compression was turned off
so I added that to both variants.

The embed variant supports a subset of protocols (http, https,
file, scp, and sftp) for now.
Most features that are built into the standard variant are turned
off.  What's left is a minimal fetch library that is intended to
be the heart of rvn fetch capability.

A specially made shell script creates a PIC library.
The zlib PIC, the SSH2 PIC and the mbedtls static libraries
need to be linked to the object using libcurl_pic.