Fix protobuf build on FreeBSD
So the problem from protobuf on FreeBSD seemed to be how clang++ was
building abseil-cpp.  I can't remember why clang was specified, but
it is not longer required apparently.

After switching to building abseil-cpp with g++, protobuf built
without issue.  It also builds with gold linker now, but I've left
it building with bfd.