guile: finally fix it's operation
Guile has been usable since the change the rvn.  Apparently this is
because rvn by default sets the file modification times to the
current time, and as a result the pre-compiled guile objects were
younger than the source, so guile wanted to rebuild the source over
and over.

By added a post-install script to touch all the ccache files,
guile finally works (and gnucash finally builds)