upgrade inkscape (poppler disabled for now)
Apparently the latest version 24.10.0 is incompatible with inkscape
and nobody has come up with a patch yet.

For now, set poppler as an off-by-default broken option and wait
for the patch (then we can turn option on by default).

Inkscape is frequently broken by poppler updates and there's no
option to use a bundled version.