cyrus-sasl: work around failure to autocreate database
I spent hours trying to make post-install operations create
the database, but I never could understand why the database
failed (since it would not fail if it was done manually).

Finally I dropped the scripts altogether in favor of
using the @sample keyword mechanism which installed a
precreated database and removes it at pkg desinstall if
unused.

Even this required a strange workaround of linking the
sasl2 subdirectory into LOCALBASE.  I think some kind
of dynamic library opening is in play.  In any case,
this @sample fix seems to work, at least on NetBSD.

While here, fix the unsorted manifest.