Drop luajit
It's now a rolling release and the developers are openly hostile to
distro maintainers that request releases.

All versions are considered incorrect (noscheme) and the other
package repositories use 2.1.<date> which is maddenly similar to
luajit-openresty.

The last 2 ports that required luajit were switched to the openresty
fork without any issue, so just drop this annoyance.