thunderbird: use bundled botan2
The botan2 port met End-Of-Line on 1 JAN 2025.
It has been removed from the tree, especially since it no longer builds
with boost.  The only port that depended on it is Thunderbird, which has
been switched to the bundled botan2 that is configured to not use boost.
When Thunderbird is updated to support botan v3, this change can be
reversed.