vlc: Several linking fixes
The libnotify subpackage needs to be removed and embedded into the primary
subpackage -- it's OS-specific.  An alterative is to have it be blank for
e.g. linux.  But as it's designed, this will fail on linux because it
will not be able to pull in libnotify:primary on linux.

while here, move execinfo and glib to various USES definitions.