Drop R-s2 and everything that depends on it
R-s2 has become extremely obnoxious to maintain.  It seems that it exists
to support R-DCluster which currently nothing uses.  I am going to assume
DCluster was orphaned and all the R-s2 nonsense is to support the orphan.
Nope.  Kill it, the orphan, and 2 other dependencies.

Obviously I can bring them back if circumstances change.