do-fetch: implement dynamic timeout
Fetch has no timeout other than the ravenadm watch which is set 1-2 hours.
The fetch executable actually has a timeout option, so let's use it.
For 99% of the downloads, a 5 minute timeout is set.
For the truly huge files (e.g. texmf) the timeouts are increased by 5m
increments up to 80 minutes.

This supports download speeds as low as 8mbps.