do-fetch.mk: new implementation of fetch timeout
The fetch -T doesn't do what I assumed it did.
It was basically setting the connection timeout which is normally
measured in a few seconds (10 -- 180 I guess).

I added timeout to all the system roots, and am using that to
send the interrupt signal if the transfer is taking too long.
If it's an older system root that lacks the timeout command, a
warning will be emitted and the original fetch command will be
used instead.