scribus: temporarily drop podofo support
Scribus doesn't build with the new version of podofo.
I started with a patch, but it got non-trivial when scribus started
using private members.  It became too much, so I decided that rather
than bring back podofo 0.10.4 as a legacy port, just turn off
podofo option and hopefully restore it in a future scribus release.