nvi: add BYTECHAR option to disable widechar
nvi was broken on OmniOS (even the solaris 2.10 version which used
to work).  After a lot of debugging, I found the L##ch macro to
convert from character strings to wide strings seemingly did not
work, which was causing the initialization of the options to fail.

My guess is that L'string' is somehow converting to UTF-8 (variable
width string) instead of a fixed-width wide string.

I need nvi working so I created an option to build NVI without
widechar functionality and NVI seems to work fine like that.

This required a new sunos patch for prototype mismatch correction
on solaris.