mesa - fix GLX
Prior to this commit, glxinfo would report:

    ~ % glxinfo
    name of display: :0
    Error: couldn't find RGB GLX visual or fbconfig

unless mesa-libs from DPorts (DragonFly ports) were found on the
system (pkg ins mesa-libs); chromium from DPorts for example depend on
this lib.

ktracing "glxinfo" revealed that it is trying to load libGLX_mesa.so.
But the latter is only build if libglvnd is found at build time, so
this commit just adds libglvnd as a build-time dependency.

GLVND is the vendor neutral interface for GL.

With this commit in place, "glxinfo" properly works.