python-ninja: hack dragonfly to avoid relinking
The recorded mtime for output files on DragonFly are just wrong.
The input file times seem to be correct.  I don't know what's going on
here.  Rather than disabling this "obsolete output" check for all platforms,
just disable it on DragonFly for now and observe.

If there are mass failures on other platforms (like FreeBSD, NetBSD) this
hack might need to be extended.