qa.mk: new check for user and group name length
This checks if user name or group name length is greater than 8.
It's debatable if this check is really needed because user names
and group names must match values in the UID and GID templates.

I guess it's good for the case invalid user name definitions
occur in the templates.

while here update valid perl interpreter values.