taskwarrior: upgrade
Even though this was a minor upgrade, the build system was significantly
changed to the point that git is required and it's trying to update
git submodules even on the downloaded distribution file.

This is a clear mistake or at least a huge failure in judgement.
Hopefully this gets fixed.  I had to patch a couple of cmake files to
avoid the use of git.