mbedtls: change source and adjust static library
We should not build from github repository.  This requires
python-jsonschema to generate a file, which requires rust to build.
This is required for rvn, and rust as a dependency is unacceptable.

Also, curl:embed couldn't link the static mbedtls without building
it's libraries with fPIC switch.  I'm not sure why this isn't a
problem for DragonFly or Linux, but adjust the compiling/linking
to make FreeBSD happy.