drop lumina-pdf port
This is the only user of poppler:qt5.
Poppler is pretty nasty in that it has to build both qt5 and qt6, making it
extremely heavy.

We have other pdf utilities, so I've decided to drop lumina-pdf so we can
remove qt5 requirements from poppler.

Poppler is the basis for *many* port so the frequent qt* rebuilds were
getting annoying.